Parsing more complete

+++ EnvironmentSimulator/Modules/CommonMini/CommonMini.cpp
@@ -1211,9 +1211,9 @@ std::string FileNameWithoutExtOf(const std::string& fname)
     }
 }
 
-bool ArrayZeroToOne(double array[], int size)
+bool ArrayZeroToOne(double array[], size_t size)
 {
-    for (int i = 0; i < size; i++)
+    for (size_t i = 0; i < size; i++)
     {
         if (array[i] > 0.0 - SMALL_NUMBER && array[i] < 1.0 + SMALL_NUMBER)
         {
